[DRIPS-562] Fix ML server startup and contract upgrade analysis endpoint
